Move VarInt and CheckedData to types module
Create a new `consensus::types` private module and put the `VarInt` and
`CheckedData` in it. These are types that exist solely for
encoding/decoding, they could go in `mod.rs` but it doesn't hurt to have
a private module just for them.
